+++ R/ebv_create_taxonomy.R
@@ -819,6 +819,15 @@ ebv_create_taxonomy <- function(jsonpath, outputpath, taxonomy, lsid=FALSE,
     ebv_i_char_att(hdf, names(global.att[i]), att.txt)
   }
 
+  #add date_modified and date_metadata_modified
+  ebv_i_char_att(hdf, 'date_modified', json$date_created)
+  ebv_i_char_att(hdf, 'date_metadata_modified', json$date_created)
+  # #add product version
+  # product_version <- stringr::str_split(stringr::str_remove(basename(jsonpath), '.json'), '_')[[1]][2]
+  # if(is.na(product_version)){
+  #   product_version <- 'v1'
+  # }
+
   #double check id - final jsons don't have 'preliminary_id' att
   id <- json$preliminary_id
   if(is.null(id)){